backend/requirements.txt
Flask
google-genai
pdfminer.six
python-docx
werkzeug
flask_cors